MOV: Exceptions to som polissa
