--- som_indexada/exceptions/__init__.py
@@ -1,1 +1,0 @@
-import indexada_exceptions
