+++ som_indexada/tests/test_wizard_change_to_periodes.py
@@ -4,7 +4,7 @@ from destral.transaction import Transaction
 from datetime import timedelta, date, datetime
 from giscedata_switching.tests.common_tests import TestSwitchingImport
 from osv import osv, fields
-from som_indexada.exceptions import indexada_exceptions as exceptions
+from som_polissa.exceptions import exceptions
 import mock
 
 class TestChangeToPeriodes(TestSwitchingImport):
